pyproject.toml
[tool.poetry]
name = "ssb-server"
version = "0.1.0"
description = ""
authors = ["Jueuunn7 <wndmsrla1234@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"
pytest-django = "^4.9.0"
django-redis = "^5.4.0"
boto3 = "^1.35.79"
django-storages = "^1.14.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
